+++ src/fullscreen/index.ts
@@ -1,3 +1,5 @@
 export * from './fullscreen.module';
 export * from './fullscreen.component';
 export * from './fullscreen.service';
+export * from './fullscreen-ref';
+export * from './fullscreen.config';